+++ src/services/ui/src/components/Form/content/FormLoadingSpinner.tsx
@@ -1,0 +1,11 @@
+import { useNavigation } from "react-router-dom";
+import { useFormContext } from "react-hook-form";
+import { LoadingSpinner } from "@/components";
+
+export const FormLoadingSpinner = () => {
+  const { state } = useNavigation();
+  const { formState } = useFormContext();
+  return (
+    (state === "submitting" || formState.isSubmitting) && <LoadingSpinner />
+  );
+};
