+++ src/services/ui/src/features/package-actions/lib/modules/issue-rai/index.tsx
@@ -10,8 +10,8 @@ import {
   ActionDescription,
   AdditionalInfoSection,
   AttachmentsSection,
+  PackageSection,
 } from "@/components";
-import { PackageSection } from "@/features/package-actions/shared-components";
 
 export const defaultIssueRaiSchema = z.object({
   additionalInformation: zAdditionalInfo,
