+++ src/services/ui/src/features/package-actions/lib/modules/respond-to-rai/spa/chip-rai.tsx
@@ -10,7 +10,7 @@ import {
   AdditionalInfoSection,
   AttachmentsSection,
 } from "@/components";
-import { PackageSection } from "@/features/package-actions/shared-components";
+import { PackageSection } from "@/components/Form/content/PackageSection";
 
 export const chipSpaRaiSchema = z.object({
   additionalInformation: zAdditionalInfo.optional(),
