+++ src/services/ui/src/features/package-actions/lib/modules/respond-to-rai/spa/med-rai.tsx
@@ -9,8 +9,8 @@ import {
   ActionDescription,
   AdditionalInfoSection,
   AttachmentsSection,
+  PackageSection,
 } from "@/components";
-import { PackageSection } from "@/features/package-actions/shared-components";
 
 export const medSpaRaiSchema = z.object({
   additionalInformation: zAdditionalInfo.optional(),
