+++ src/services/ui/src/features/package-actions/lib/modules/respond-to-rai/waiver/b-waiver-rai.tsx
@@ -6,7 +6,7 @@ import {
   AdditionalInfoSection,
   AttachmentsSection,
 } from "@/components";
-import { PackageSection } from "@/features/package-actions/shared-components";
+import { PackageSection } from "@/components/Form/content/PackageSection";
 
 export const bWaiverRaiSchema = z.object({
   additionalInformation: z.string().optional().default(""),
