+++ src/services/ui/src/features/package-actions/lib/modules/withdraw-rai/index.tsx
@@ -6,8 +6,8 @@ import {
   ActionDescription,
   AdditionalInfoSection,
   AttachmentsSection,
+  PackageSection,
 } from "@/components";
-import { PackageSection } from "@/features/package-actions/shared-components";
 
 export const defaultWithdrawRaiSchema = z.object({
   additionalInformation: zAdditionalInfo,
