--- src/services/ui/src/features/package-actions/shared-components.tsx
@@ -1,310 +1,0 @@
-import {
-  Alert,
-  AttachmentsSizeTypesDesc,
-  LoadingSpinner,
-  Route,
-  SubmissionAlert,
-  useAlertContext,
-  useModalContext,
-} from "@/components";
-import {
-  Button,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-  RequiredIndicator,
-  Textarea,
-  Upload,
-} from "@/components/Inputs";
-import { FAQ_TAB } from "@/components/Routing/consts";
-import { useEffect } from "react";
-import { Path, SubmitHandler, useFormContext } from "react-hook-form";
-import {
-  ActionFunctionArgs,
-  Link,
-  useActionData,
-  useLocation,
-  useNavigate,
-  useNavigation,
-  useParams,
-  useSubmit,
-} from "react-router-dom";
-import { Authority } from "shared-types";
-import { AttachmentRecipe } from "@/utils";
-
-// Components
-
-export const RequiredFieldDescription = () => {
-  return (
-    <>
-      <span className="text-red-500">*</span> Indicates a required field
-    </>
-  );
-};
-
-export const ActionDescription = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  return <div className="font-light mb-6 max-w-4xl">{children}</div>;
-};
-
-export const Heading = ({ title }: { title: string }) => {
-  return <h1 className="text-2xl font-semibold mt-4 mb-2">{title}</h1>;
-};
-
-export const AttachmentsSection = ({
-  attachments,
-  instructions,
-  faqLink,
-}: {
-  attachments: AttachmentRecipe<any>[];
-  instructions?: string;
-  faqLink: string;
-}) => {
-  const form = useFormContext();
-  return (
-    <>
-      <h2 className="font-bold text-2xl font-sans mb-2">Attachments</h2>
-      {instructions && <p>{instructions}</p>}
-      <AttachmentsSizeTypesDesc faqLink={faqLink} />
-      {attachments.map(({ name, label, required }) => (
-        <FormField
-          key={String(name) + "-field"}
-          control={form.control}
-          name={`attachments.${String(name)}`}
-          render={({ field }) => (
-            <FormItem key={String(name) + "-render"} className="my-4 space-y-2">
-              <FormLabel>{label}</FormLabel> {required && <RequiredIndicator />}
-              <Upload files={field?.value ?? []} setFiles={field.onChange} />
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-      ))}
-    </>
-  );
-};
-
-/** Action form submission buttons. Use the `confirmWithdraw` prop when performing
- * a destructive action like a withdrawal. */
-export const SubmissionButtons = ({
-  confirmWithdraw,
-}: {
-  confirmWithdraw?: () => void;
-}) => {
-  const { state } = useNavigation();
-  const modal = useModalContext();
-  const navigate = useNavigate();
-
-  const acceptAction = () => {
-    modal.setModalOpen(false);
-    navigate(-1);
-  };
-
-  return (
-    <section className="space-x-2 mb-8">
-      {confirmWithdraw && (
-        <Button
-          type={"button"}
-          onClick={() => confirmWithdraw()}
-          disabled={state === "submitting"}
-        >
-          Submit
-        </Button>
-      )}
-      {!confirmWithdraw && (
-        <Button type={"submit"} disabled={state === "submitting"}>
-          Submit
-        </Button>
-      )}
-      <Button
-        onClick={() => {
-          modal.setContent({
-            header: "Stop form submission?",
-            body: "All information you've entered on this form will be lost if you leave this page.",
-            acceptButtonText: "Yes, leave form",
-            cancelButtonText: "Return to form",
-          });
-          modal.setOnAccept(() => acceptAction);
-          modal.setModalOpen(true);
-        }}
-        variant={"outline"}
-        type="reset"
-        disabled={state === "submitting"}
-      >
-        Cancel
-      </Button>
-    </section>
-  );
-};
-
-export const PackageSection = () => {
-  const { authority, id } = useParams() as { authority: Authority; id: string };
-  const lcAuthority = authority.toLowerCase();
-  // We should pass in the already lowercased Authority, right?  todo
-  return (
-    <section className="flex flex-col my-8 space-y-8">
-      <div>
-        <p>
-          {[Authority.CHIP_SPA, Authority.MED_SPA].includes(
-            authority.toLowerCase() as Authority,
-          ) && "Package ID"}
-          {[Authority["1915b"], Authority["1915c"]].includes(
-            authority.toLowerCase() as Authority,
-          ) && "Waiver Number"}
-        </p>
-        <p className="text-xl">{id}</p>
-      </div>
-      <div>
-        <p>Authority</p>
-        <p className="text-xl">
-          {lcAuthority === Authority["1915b"] && "1915(b) Waiver"}
-          {lcAuthority === Authority["1915c"] && "1915(c) Waiver"}
-          {lcAuthority === Authority["CHIP_SPA"] && "CHIP SPA"}
-          {lcAuthority === Authority["MED_SPA"] && "Medicaid SPA"}
-        </p>
-      </div>
-    </section>
-  );
-};
-
-export const ErrorBanner = () => {
-  const form = useFormContext();
-  return Object.keys(form.formState.errors).length !== 0 ? (
-    <Alert className="my-6" variant="destructive">
-      Input validation error(s)
-      <ul className="list-disc">
-        {Object.values(form.formState.errors).map(
-          (err, idx) =>
-            err?.message && (
-              <li className="ml-8 my-2" key={idx}>
-                {err.message as string}
-              </li>
-            ),
-        )}
-      </ul>
-    </Alert>
-  ) : null;
-};
-
-export const FormLoadingSpinner = () => {
-  const { state } = useNavigation();
-  const { formState } = useFormContext();
-  return (
-    (state === "submitting" || formState.isSubmitting) && <LoadingSpinner />
-  );
-};
-
-// Hooks
-
-export const useSubmitForm = () => {
-  const methods = useFormContext();
-  const submit = useSubmit();
-  const location = useLocation();
-
-  const validSubmission: SubmitHandler<any> = (data, e) => {
-    const formData = new FormData();
-    // Append all other data
-    for (const key in data) {
-      if (key !== "attachments") {
-        formData.append(key, data[key]);
-      }
-    }
-    const attachments =
-      Object.keys(filterUndefinedValues(data.attachments)).length > 0
-        ? data.attachments
-        : {};
-    for (const key in attachments) {
-      attachments[key]?.forEach((file: any, index: number) => {
-        formData.append(`attachments.${key}.${index}`, file as any);
-      });
-    }
-
-    submit(formData, {
-      method: "post",
-      encType: "multipart/form-data",
-      state: location.state,
-    });
-  };
-
-  return {
-    handleSubmit: methods.handleSubmit(validSubmission),
-    formMethods: methods,
-  };
-};
-
-export const useIntakePackage = () => {
-  const methods = useFormContext();
-  const submit = useSubmit();
-  const location = useLocation();
-
-  const validSubmission: SubmitHandler<any> = (data, e) => {
-    submit(data, {
-      method: "post",
-      encType: "application/json",
-      state: location.state,
-    });
-  };
-
-  return {
-    handleSubmit: methods.handleSubmit(validSubmission),
-    formMethods: methods,
-  };
-};
-
-export const useDisplaySubmissionAlert = (header: string, body: string) => {
-  const alert = useAlertContext();
-  const data = useActionData() as ActionFunctionReturnType;
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  return useEffect(() => {
-    if (data?.submitted) {
-      alert.setContent({
-        header,
-        body,
-      });
-      alert.setBannerStyle("success");
-      alert.setBannerShow(true);
-      if (location.pathname?.endsWith("/update-id")) {
-        alert.setBannerDisplayOn("/dashboard");
-        navigate("/dashboard");
-      } else {
-        alert.setBannerDisplayOn(
-          location.state?.from?.split("?")[0] ?? "/dashboard",
-        );
-        navigate(location.state?.from ?? "/dashboard");
-      }
-    } else if (!data?.submitted && data?.error) {
-      alert.setContent({
-        header: "An unexpected error has occurred:",
-        body:
-          data.error instanceof Error ? data.error.message : String(data.error),
-      });
-      alert.setBannerStyle("destructive");
-      alert.setBannerDisplayOn(window.location.pathname as Route);
-      alert.setBannerShow(true);
-      window.scrollTo(0, 0);
-    }
-  }, [data, alert, navigate, location.state, location.pathname]);
-};
-
-// Utility Functions
-const filterUndefinedValues = (obj: Record<any, any>) => {
-  if (obj) {
-    return Object.fromEntries(
-      Object.entries(obj).filter(([key, value]) => value !== undefined),
-    );
-  }
-  return {};
-};
-
-// Types
-export type ActionFunction = (
-  args: ActionFunctionArgs,
-) => Promise<{ submitted: boolean; error?: Error | unknown }>;
-export type ActionFunctionReturnType = Awaited<ReturnType<ActionFunction>>;
